requirements.txt
click==8.1.7
Flask==2.3.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
python-dotenv==1.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
greenlet==2.0.2
